editing orders or downloads connection parametrs is updated with action data sources
conditions are updated with 'add_action' values.

request are also checking for 'actions_orders_edit' and 'actions_downloads_edit'.

the request must have process = connected argument (if data source connection is not failed). This is because both customers and sessions of active connections are created individually. However, Main or Dimensional data for Sessions and Customers must be crated at the same time for the same ElasticSerach server.